renamed class dhcpv6snoop, as never use the same name for a class and a namespace

+++ src/dhcpv6snoop/cdhcp_option_ia_pd_option_prefix.h
@@ -93,9 +93,9 @@ public:
 	operator<< (std::ostream& os, const cdhcp_option_ia_pd_option_prefix& opt) {
 		os << dynamic_cast<const cdhcp_option&>( opt );
 		os << "<dhcp-option-ia-pd-option-prefix: ";
-			os << "preferred-lifetime: " << opt.get_preferred_lifetime() << " ";
-			os << "valid-lifetime: " << opt.get_valid_lifetime() << " ";
-			os << "prefixlen: " << opt.get_prefixlen() << " ";
+			os << "preferred-lifetime: " << (unsigned int)opt.get_preferred_lifetime() << " ";
+			os << "valid-lifetime: " << (unsigned int)opt.get_valid_lifetime() << " ";
+			os << "prefixlen: " << (unsigned int)opt.get_prefixlen() << " ";
 			os << "prefix: " << opt.get_prefix() << " ";
 		os << ">";
 		return os;
